[Bug#2] Fix Product Delivery with Charge
The previous version of the script can only handle products with FREE delivery cost. Now it can also handle if any products has delivery cost.

If the delivery cost is not free. The script will split the striped text from web and slips the first index(0) of the string(the first index has the cost)